+++ core/src/main/java/me/mykindos/betterpvp/core/client/stats/events/IStatMapListener.java
@@ -4,5 +4,5 @@ import me.mykindos.betterpvp.core.client.stats.impl.IStat;
 import org.jetbrains.annotations.Nullable;
 
 public interface IStatMapListener {
-    void onMapValueChanged(IStat stat, Double newValue, @Nullable Double oldValue);
+    void onMapValueChanged(IStat stat, Long newValue, @Nullable Long oldValue);
 }
